[kbss-cvut/termit#164] Introduce a dedicated module for authorization.
Move isLoggedIn and isEditor functions to it from SecurityUtils.
