+++ src/component/vocabulary/VocabularyIriLink.tsx
@@ -6,8 +6,8 @@ import AssetIriLink from "../misc/AssetIriLink";
 import User from "../../model/User";
 import { connect } from "react-redux";
 import TermItState from "../../model/TermItState";
-import SecurityUtils from "../../util/SecurityUtils";
 import { useI18n } from "../hook/useI18n";
+import { isLoggedIn } from "../../util/Authorization";
 
 interface VocabularyIriLinkProps {
   iri: string;
@@ -27,7 +27,7 @@ export const VocabularyIriLink: React.FC<VocabularyIriLinkProps> = (
   const { i18n } = useI18n();
   const iri = VocabularyUtils.create(props.iri);
   const path = Routing.getTransitionPath(
-    SecurityUtils.isLoggedIn(props.user)
+    isLoggedIn(props.user)
       ? Routes.vocabularySummary
       : Routes.publicVocabularySummary,
     {
