+++ src/util/__tests__/SecurityUtils.test.ts
@@ -1,9 +1,6 @@
 import SecurityUtils from "../SecurityUtils";
 import Constants from "../Constants";
-import { EMPTY_USER } from "../../model/User";
-import Generator from "../../__tests__/environment/Generator";
 import BrowserStorage from "../BrowserStorage";
-import VocabularyUtils from "../VocabularyUtils";
 
 jest.mock("../BrowserStorage");
 
@@ -37,39 +34,5 @@ describe("SecurityUtils", () => {
     expect(BrowserStorage.remove).toHaveBeenCalledWith(
       Constants.STORAGE_JWT_KEY
     );
-  });
-
-  describe("isLoggedIn", () => {
-    it("returns false for empty user", () => {
-      expect(SecurityUtils.isLoggedIn(null)).toBeFalsy();
-      expect(SecurityUtils.isLoggedIn(undefined)).toBeFalsy();
-      expect(SecurityUtils.isLoggedIn(EMPTY_USER)).toBeFalsy();
-    });
-
-    it("returns true for non-empty user", () => {
-      expect(SecurityUtils.isLoggedIn(Generator.generateUser())).toBeTruthy();
-    });
-  });
-
-  describe("isEditor", () => {
-    it("returns false when user has restricted user type", () => {
-      const user = Generator.generateUser();
-      user.types.push(VocabularyUtils.USER_RESTRICTED);
-      expect(SecurityUtils.isEditor(user)).toBeFalsy();
-    });
-
-    it("returns true for regular user and admin", () => {
-      let user = Generator.generateUser();
-      user.types.push(VocabularyUtils.USER_EDITOR);
-      expect(SecurityUtils.isEditor(user)).toBeTruthy();
-      user = Generator.generateUser();
-      user.types.push(VocabularyUtils.USER_ADMIN);
-      expect(SecurityUtils.isEditor(user)).toBeTruthy();
-    });
-
-    it("returns false when user is not logged in at all", () => {
-      expect(SecurityUtils.isEditor(null)).toBeFalsy();
-      expect(SecurityUtils.isEditor(undefined)).toBeFalsy();
-    });
   });
 });
